+++ sale-page-et2-maxim-update.php
@@ -39,16 +39,16 @@
                         <div class="block-picture-primary-container">
 
                             <div class="block-container">
-                                <a href="<?php echo site_url('/brands/quoizel-lighting/'); ?>">
+                                <a href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-table-lamps/'); ?>">
                                     <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/04/LWN5005BN_4_.jpg">', wp_get_attachment_metadata(15942), 15942 ); ?>
+                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/05/TF6668VB.jpg">', wp_get_attachment_metadata(22354), 22354 ); ?>
                                     </div>
                                     <div class="block-caption-wrapper">
                                         <div class="block-caption-container">
-                                            <h4 class="block-heading">Quoizel Lighting</h4>
-                                            <h3 class="block-heading">Get 10% off on All of Quoizel Lighting</h3>
-                                            <h4 class="block-heading">Use code "QUOIZELALL10" at checkout!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/quoizel-lighting/'); ?>" class="block-shop-link">Shop Now</a>
+                                            <h4 class="block-heading">Quoizel Table Lamps</h4>
+                                            <h3 class="block-heading">Get 10% off on Quoizel Table Lamps</h3>
+                                            <h4 class="block-heading">Use code "QUOIZEL10" at checkout!</h4>
+                                            <a style="text-align: center;" href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-table-lamps/'); ?>" class="block-shop-link">Shop Now</a>
                                         </div>
                                     </div>
                                 </a>
@@ -58,25 +58,49 @@
 
                     </div>
 
-                    <!-- <div class="col-md-4">
+                    <div class="col-md-4">
+
+                        <div class="block-picture-primary-container">
+
+                            <div class="block-container">
+                                <a href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-floor-lamps/'); ?>">
+                                    <div class="block-picture">
+                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/05/TFMK9362VA.jpg">', wp_get_attachment_metadata(22438), 22438 ); ?>
+                                    </div>
+                                    <div class="block-caption-wrapper">
+                                        <div class="block-caption-container">
+                                            <h4 class="block-heading">Quoizel Floor Lamps</h4>
+                                            <h3 class="block-heading">Get 10% off on Quoizel Floor Lamps</h3>
+                                            <h4 class="block-heading">Use code "QUOIZEL10" at checkout!</h4>
+                                            <a style="text-align: center;" href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-floor-lamps/'); ?>" class="block-shop-link">Shop Now</a>
+                                        </div>
+                                    </div>
+                                </a>
+                            </div>
+
+                        </div>
+
+                    </div>
+
+                    <div class="col-md-4">
 
                         <div class="block-container">
-                            <a href="<?php //echo site_url('/brands/elk-home'); ?>">
+                            <a href="<?php echo site_url('/brands/elk-home'); ?>">
                                 <div class="block-picture">
-                                    <?php //echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/07/elk_home_sale_august.jpg">', wp_get_attachment_metadata(172838), 172838 ); ?>
+                                    <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/07/elk_home_sale_august.jpg">', wp_get_attachment_metadata(172838), 172838 ); ?>
                                 </div>
                                 <div class="block-caption-wrapper">
                                     <div class="block-caption-container">
                                         <h4 class="block-heading">Elk Home</h4>
                                         <h3 class="block-heading">Save 15% on all Elk Home Products</h3>
                                         <h4 class="block-heading">Use code ELKHOME15 at checkout!</h4>
-                                        <a style="text-align: center;" href="<?php //echo site_url('/brands/elk-home'); ?>" class="block-shop-link">Shop Now</a>
+                                        <a style="text-align: center;" href="<?php echo site_url('/brands/elk-home'); ?>" class="block-shop-link">Shop Now</a>
                                     </div>
                                 </div>
                              </a>
                         </div>
 
-                    </div> -->
+                    </div>
 
                     <div class="col-md-4">
 
@@ -102,41 +126,25 @@
 
                     </div>
 
-                    <!-- <div class="col-md-4">
+                    <div class="col-md-4">
                         <div class="block-container">
-                            <a href="<?php //echo site_url('/brands/maxim-lighting'); ?>">
+                            <a href="<?php echo site_url('https://www.lightyourhome.com/product-category/maxim-lighting/maxim-lighting-outdoor-lighting/'); ?>">
                                 <div class="block-picture">
-                                     <?php //echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/04/31213OIAB-scaled.jpg">', wp_get_attachment_metadata(165335), 165335 ); ?>
+                                     <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/04/10316CLBZGLD.jpg">', wp_get_attachment_metadata(165421), 165421 ); ?>
                                 </div>
                                 <div class="block-caption-wrapper">
                                     <div class="block-caption-container">
                                         <h4 class="block-heading">Maxim Lighting</h4>
-                                        <h3 class="block-heading">Save 10% on All Maxim Lighting 2019-2020 Releases</h3>
+                                        <h3 class="block-heading">Save 10% on All Maxim Outdoor Lighting</h3>
                                         <h4 class="block-heading">Prices already discounted!</h4>
-                                        <a style="text-align: center;" href="<?php //echo site_url('/brands/maxim-lighting'); ?>" class="block-shop-link">Shop Now</a>
+                                        <a style="text-align: center;" href="<?php echo site_url('https://www.lightyourhome.com/product-category/maxim-lighting/maxim-lighting-outdoor-lighting/'); ?>" class="block-shop-link">Shop Now</a>
                                     </div>
                                 </div>
                             </a>
                          </div>
-                    </div> -->
+                    </div>
 
-                     <div class="col-md-4">
-                        <div class="block-container">
-                            <a href="<?php echo site_url('/brands/et2-lighting'); ?>">
-                                <div class="block-picture">
-                                     <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/04/E25064-92BKGLD-scaled.jpg">', wp_get_attachment_metadata(165986), 165986 ); ?>
-                                </div>
-                                <div class="block-caption-wrapper">
-                                    <div class="block-caption-container">
-                                        <h4 class="block-heading">ET2 Lighting</h4>
-                                        <h3 class="block-heading">Save 10% on All ET2 Lighting 2019-2020 Releases</h3>
-                                        <h4 class="block-heading">Prices already discounted!</h4>
-                                        <a style="text-align: center;" href="<?php echo site_url('/brands/et2-lighting'); ?>" class="block-shop-link">Shop Now</a>
-                                    </div>
-                                </div>
-                            </a>
-                         </div>
-                     </div>
+                     
 
                     <div class="col-md-12" style="border-top: 2px solid #d3d3d3;"></div>