+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable theme changes should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [ unreleased ] - scheduled for 10.1.2020
+
+## [ 1.21.0 ]
+
+## Added
+
+- Elk Home sale to home page
+- Outdoor Wall Lighting to Maxim brand page template
+- Elk Home sale to sale page
+- Quoizel Table and floors lamps sale to sale page
+
+
+## Changed
+
+- Maxim Lighting Sale on sale page
+- Single product page coupon codes
+
+## Removed
+
+- Quoizel sale from home page
+- Removed ET2 sale from sale page
+- Quoizel All Sale from sale page
+
 ## [ released ] - 9.30.2020
 
 ## [ 1.20.2 ]